Move 'Feed Presentation' module production code to their own files from 'FeedViewPresenterTests'
