+++ 3.EssentialFeedProject/EssentialFeedProject/EssentialFeed/EssentialFeed/Feed Presentation/FeedViewPresenter.swift
@@ -1,0 +1,46 @@
+//
+//  FeedViewPresenter.swift
+//  EssentialFeed
+//
+//  Created by Fardan Akhter on 13/05/2023.
+//
+
+import Foundation
+
+public protocol FeedView {
+    func display(_ viewModel: FeedViewModel)
+}
+
+public protocol FeedLoadingView {
+    func display(_ viewModel: FeedLoadingViewModel)
+}
+
+public final class FeedViewPresenter {
+    private let feedView: FeedView
+    private let loadingView: FeedLoadingView
+    
+    public init(feedView: FeedView, loadingView: FeedLoadingView) {
+        self.feedView = feedView
+        self.loadingView = loadingView
+    }
+    
+   public static var feedViewTitle: String {
+        NSLocalizedString("FEED_VIEW_TITLE",
+                          tableName: "Feed",
+                          bundle: Bundle(for: FeedViewPresenter.self),
+                          comment: "Title for Feed View")
+    }
+    
+    public func didStartLoadingFeed() {
+        loadingView.display(FeedLoadingViewModel(isLoading: true))
+    }
+    
+    public func didCompleteLoadingFeed(with feed: [FeedImage]) {
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+        feedView.display(FeedViewModel(feed: feed))
+    }
+    
+    public func didCompleteLoadingFeed(with error: Error) {
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+}
