+++ 3.EssentialFeedProject/EssentialFeedProject/EssentialFeed/EssentialFeed/Feed Presentation/Models/FeedLoadingViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  FeedLoadingViewModel.swift
+//  EssentialFeed
+//
+//  Created by Fardan Akhter on 13/05/2023.
+//
+
+import Foundation
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
