+++ 3.EssentialFeedProject/EssentialFeedProject/EssentialFeed/EssentialFeed/Feed Presentation/Models/FeedViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  ViewModel.swift
+//  EssentialFeed
+//
+//  Created by Fardan Akhter on 13/05/2023.
+//
+
+import Foundation
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
+
